TW-816: refactor new group chat and new group chat info screen
