+++ lib/pages/new_group/new_group.dart
@@ -1,35 +1,8 @@
-import 'dart:async';
-
-import 'package:collection/collection.dart';
-import 'package:dartz/dartz.dart' hide State;
-import 'package:file_picker/file_picker.dart';
-import 'package:fluffychat/app_state/failure.dart';
-import 'package:fluffychat/app_state/success.dart';
-import 'package:fluffychat/di/global/get_it_initializer.dart';
-import 'package:fluffychat/domain/app_state/room/create_new_group_chat_state.dart';
-import 'package:fluffychat/domain/app_state/room/upload_content_state.dart';
-import 'package:fluffychat/domain/model/room/create_new_group_chat_request.dart';
-import 'package:fluffychat/domain/usecase/room/create_new_group_chat_interactor.dart';
-import 'package:fluffychat/domain/usecase/room/upload_content_interactor.dart';
-import 'package:fluffychat/domain/usecase/room/upload_content_for_web_interactor.dart';
 import 'package:fluffychat/pages/new_group/contacts_selection.dart';
-import 'package:fluffychat/pages/new_group/contacts_selection_view.dart';
-import 'package:fluffychat/pages/new_group/new_group_chat_info.dart';
-import 'package:fluffychat/pages/new_group/new_group_info_controller.dart';
-import 'package:fluffychat/presentation/mixins/common_media_picker_mixin.dart';
-import 'package:fluffychat/presentation/mixins/single_image_picker_mixin.dart';
-import 'package:fluffychat/presentation/model/chat/chat_router_input_argument.dart';
-import 'package:fluffychat/presentation/model/presentation_contact.dart';
-import 'package:fluffychat/utils/dialog/warning_dialog.dart';
 import 'package:fluffychat/utils/platform_infos.dart';
-import 'package:fluffychat/widgets/matrix.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/l10n.dart';
 import 'package:go_router/go_router.dart';
-import 'package:linagora_design_flutter/images_picker/asset_counter.dart';
-import 'package:linagora_design_flutter/images_picker/images_picker.dart';
-import 'package:matrix/matrix.dart';
-import 'package:wechat_camera_picker/wechat_camera_picker.dart';
 
 class NewGroup extends StatefulWidget {
   const NewGroup({Key? key}) : super(key: key);
@@ -38,40 +11,7 @@ class NewGroup extends StatefulWidget {
   NewGroupController createState() => NewGroupController();
 }
 
-class NewGroupController extends ContactsSelectionController<NewGroup>
-    with CommonMediaPickerMixin, SingleImagePickerMixin {
-  final uploadContentInteractor = getIt.get<UploadContentInteractor>();
-  final uploadContentWebInteractor =
-      getIt.get<UploadContentInBytesInteractor>();
-  final createNewGroupChatInteractor =
-      getIt.get<CreateNewGroupChatInteractor>();
-  final groupNameTextEditingController = TextEditingController();
-  final avatarAssetEntityNotifier = ValueNotifier<AssetEntity?>(null);
-  final avatarFilePickerNotifier = ValueNotifier<FilePickerResult?>(null);
-  final createRoomStateNotifier =
-      ValueNotifier<Either<Failure, Success>>(Right(CreateNewGroupInitial()));
-
-  final haveGroupNameNotifier = ValueNotifier(false);
-  final groupNameFocusNode = FocusNode();
-  StreamSubscription? createNewGroupChatInteractorStreamSubscription;
-  final enableEncryptionNotifier = ValueNotifier(false);
-
-  String groupName = "";
-
-  @override
-  void initState() {
-    super.initState();
-    listenGroupNameChanged();
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-    haveGroupNameNotifier.dispose();
-    avatarAssetEntityNotifier.dispose();
-    avatarFilePickerNotifier.dispose();
-    createNewGroupChatInteractorStreamSubscription?.cancel();
-  }
+class NewGroupController extends ContactsSelectionController<NewGroup> {
 
   @override
   String getTitle(BuildContext context) {
@@ -88,295 +28,12 @@ class NewGroupController extends ContactsSelectionController<NewGroup>
     moveToNewGroupInfoScreen();
   }
 
-  Future<ServerConfig> getServerConfig() async {
-    final serverConfig = await Matrix.of(context).client.getConfig();
-    return serverConfig;
-  }
-
-  Future<Set<PresentationContact>> getAllContactsGroupChat({
-    bool isCustomDisplayName = true,
-  }) async {
-    final userId = Matrix.of(context).client.userID;
-    final profile =
-        await Matrix.of(context).client.getProfileFromUserId(userId!);
-    final newContactsList = {
-      PresentationContact(
-        displayName:
-            isCustomDisplayName ? L10n.of(context)!.you : profile.displayName,
-        matrixId: Matrix.of(context).client.userID,
-      ),
-    };
-    newContactsList.addAll(getSelectedValidContacts(contactsList));
-    return newContactsList;
-  }
-
-  void _getDefaultGroupName(Set<PresentationContact> contactLis) async {
-    if (contactLis.length <= 3) {
-      final groupName =
-          contactLis.map((contact) => contact.displayName).join(", ");
-      groupNameTextEditingController.text = groupName;
-      groupNameTextEditingController.selection = TextSelection.fromPosition(
-        TextPosition(offset: groupNameTextEditingController.text.length),
-      );
-      groupNameFocusNode.requestFocus();
-    } else {
-      groupNameTextEditingController.clear();
-    }
-  }
-
   void moveToNewGroupInfoScreen() async {
-    final contactList =
-        await getAllContactsGroupChat(isCustomDisplayName: false);
-    _getDefaultGroupName(contactList);
-    await showGeneralDialog(
-      pageBuilder: (context, animation, secondaryAnimation) {
-        return const SizedBox.shrink();
-      },
-      context: context,
-      useRootNavigator: false,
-      barrierColor: Colors.white,
-      transitionBuilder: (context, animation1, animation2, widget) {
-        return NewGroupChatInfo(
-          contactsList: contactList,
-          newGroupController: this,
-        );
-      },
-    );
-  }
-
-  void createNewGroup({String? urlAvatar}) {
-    final client = Matrix.of(context).client;
-    createNewGroupChatAction(
-      matrixClient: client,
-      createNewGroupChatRequest: CreateNewGroupChatRequest(
-        groupName: groupName,
-        invite: getSelectedValidContacts(contactsList)
-            .map((contact) => contact.matrixId)
-            .whereNotNull()
-            .toList(),
-        enableEncryption: enableEncryptionNotifier.value,
-        urlAvatar: urlAvatar,
-      ),
-    );
-  }
-
-  void uploadAvatarNewGroupChat({
-    required Client matrixClient,
-    required AssetEntity entity,
-  }) {
-    uploadContentInteractor
-        .execute(
-          matrixClient: matrixClient,
-          entity: entity,
-        )
-        .listen(
-          (event) => _handleUploadAvatarNewGroupChatOnData(context, event),
-          onDone: _handleUploadAvatarNewGroupChatOnDone,
-          onError: _handleUploadAvatarNewGroupChatOnError,
-        );
-  }
-
-  void uploadAvatarNewGroupChatInBytes({
-    required Client matrixClient,
-    required FilePickerResult filePickerResult,
-  }) {
-    uploadContentWebInteractor
-        .execute(
-          matrixClient: matrixClient,
-          filePickerResult: filePickerResult,
-        )
-        .listen(
-          (event) => _handleUploadAvatarNewGroupChatOnData(context, event),
-          onDone: _handleUploadAvatarNewGroupChatOnDone,
-          onError: _handleUploadAvatarNewGroupChatOnError,
-        );
-  }
-
-  void createNewGroupChatAction({
-    required Client matrixClient,
-    required CreateNewGroupChatRequest createNewGroupChatRequest,
-  }) {
-    createNewGroupChatInteractorStreamSubscription =
-        createNewGroupChatInteractor
-            .execute(
-              matrixClient: matrixClient,
-              createNewGroupChatRequest: createNewGroupChatRequest,
-            )
-            .listen(
-              (event) => _handleCreateNewGroupChatChatOnData(context, event),
-              onDone: _handleCreateNewGroupChatOnDone,
-              onError: _handleCreateNewGroupChatOnError,
-            );
-  }
-
-  void _handleUploadAvatarNewGroupChatOnData(
-    BuildContext context,
-    Either<Failure, Success> event,
-  ) {
-    Logs().d('NewGroupController::_handleUploadAvatarNewGroupChatOnData()');
-    createRoomStateNotifier.value = event;
-    event.fold(
-      (failure) {
-        Logs().e(
-          'NewGroupController::_handleUploadAvatarNewGroupChatOnData() - failure: $failure',
-        );
-        WarningDialog.showCancelable(
-          context,
-          message: L10n.of(context)!
-              .youAreUploadingPhotosDoYouWantToCancelOrContinue,
-          acceptText: L10n.of(context)!.continueProcess,
-          onAccept: () {
-            WarningDialog.hideWarningDialog(context);
-            createNewGroup();
-          },
-        );
-      },
-      (success) {
-        Logs().d(
-          'NewGroupController::_handleUploadAvatarNewGroupChatOnData() - success: $success',
-        );
-        if (success is UploadContentSuccess) {
-          final urlAvatar = success.uri.toString();
-          createNewGroup(urlAvatar: urlAvatar);
-        }
-      },
-    );
-  }
-
-  void _handleUploadAvatarNewGroupChatOnDone() {
-    Logs().d(
-      'NewGroupController::_handleUploadAvatarNewGroupChatOnDone() - done',
-    );
-  }
-
-  void _handleUploadAvatarNewGroupChatOnError(
-    dynamic error,
-    StackTrace? stackTrace,
-  ) {
-    Logs().e(
-      'NewGroupController::_handleUploadAvatarNewGroupChatOnError() - error: $error | stackTrace: $stackTrace',
-    );
-  }
-
-  void _handleCreateNewGroupChatChatOnData(
-    BuildContext context,
-    Either<Failure, Success> event,
-  ) {
-    Logs().d('NewGroupController::_handleCreateNewGroupChatChatOnData()');
-    createRoomStateNotifier.value = event;
-    event.fold(
-      (failure) {
-        Logs().e(
-          'NewGroupController::_handleCreateNewGroupChatChatOnData() - failure: $failure',
-        );
-      },
-      (success) {
-        Logs().d(
-          'NewGroupController::_handleCreateNewGroupChatChatOnData() - success: $success',
-        );
-        if (success is CreateNewGroupChatSuccess) {
-          _goToRoom(success);
-        }
-      },
-    );
-  }
-
-  void _handleCreateNewGroupChatOnDone() {
-    Logs().d('NewGroupController::_handleCreateNewGroupChatOnDone() - done');
-  }
-
-  void _handleCreateNewGroupChatOnError(dynamic error, StackTrace? stackTrace) {
-    Logs().e(
-      'NewGroupController::_handleUploadAvatarNewGroupChatOnError() - error: $error | stackTrace: $stackTrace',
-    );
-  }
-
-  void _goToRoom(CreateNewGroupChatSuccess success) {
-    context.go(
-      "/rooms/${success.roomId}",
-      extra: ChatRouterInputArgument(
-        type: ChatRouterInputArgumentType.draft,
-        data: success.groupName,
-      ),
-    );
-  }
-
-  void _getImageOnWeb(
-    BuildContext context,
-  ) async {
-    final result = await FilePicker.platform.pickFiles(
-      type: FileType.image,
-    );
-    Logs().d(
-      'NewGroupController::_getImageOnWeb(): FilePickerResult - $result',
-    );
-    if (result == null || result.files.single.bytes == null) {
-      return;
-    } else {
-      avatarFilePickerNotifier.value = result;
-      Logs().d(
-        'NewGroupController::_getImageOnWeb(): AvatarWebNotifier - ${avatarFilePickerNotifier.value}',
+    if (PlatformInfos.isMobile) {
+      context.push(
+        '/rooms/newprivatechat/newgroup/newgroupinfo',
+        extra: contactsList.toSet(),
       );
-    }
+    } else {}
   }
-
-  void showImagesPickerAction({
-    required BuildContext context,
-  }) async {
-    if (isCreatingRoom) {
-      return;
-    }
-    if (PlatformInfos.isWeb) {
-      _getImageOnWeb(context);
-      return;
-    }
-    final currentPermissionPhotos = await getCurrentMediaPermission();
-    final currentPermissionCamera = await getCurrentCameraPermission();
-    if (currentPermissionPhotos != null && currentPermissionCamera != null) {
-      final imagePickerController = createImagePickerController();
-      groupNameFocusNode.unfocus();
-      showImagePickerBottomSheet(
-        context,
-        currentPermissionPhotos,
-        currentPermissionCamera,
-        imagePickerController,
-      );
-    }
-  }
-
-  ImagePickerGridController createImagePickerController() {
-    final imagePickerController = ImagePickerGridController(
-      AssetCounter(imagePickerMode: ImagePickerMode.single),
-    );
-
-    imagePickerController.addListener(() {
-      final selectedAsset = imagePickerController.selectedAssets.firstOrNull;
-      if (selectedAsset?.asset.type == AssetType.image) {
-        if (!imagePickerController.pickFromCamera()) {
-          Navigator.pop(context);
-        }
-        avatarAssetEntityNotifier.value = selectedAsset?.asset;
-        imagePickerController.removeAllSelectedItem();
-      }
-    });
-
-    return imagePickerController;
-  }
-
-  bool get isCreatingRoom {
-    return createRoomStateNotifier.value.fold(
-          (failure) => false,
-          (success) =>
-              success is UploadContentLoading ||
-              success is CreateNewGroupChatLoading,
-        ) ??
-        false;
-  }
-
-  void toggleEnableEncryption() {
-    enableEncryptionNotifier.value = !enableEncryptionNotifier.value;
-  }
-
-  @override
-  Widget build(BuildContext context) => ContactsSelectionView(this);
 }
